# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAtext API
  description: REST API specification for WASAtext messaging system.
  version: "3.0.1"
servers:
  - url: http://localhost:3000
    description: Test server running locally, for development purposes.
...
paths:
...
  /me/photo:
    get:
      tags: [profile]
      summary: Get my profile photo
      description: Retrieves the authenticated user's profile photo.
      operationId: getMyPhoto
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Profile photo retrieved successfully
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
        "401":
          description: Unauthorized - invalid or missing token
        "404":
          description: Photo not found
        "500":
          description: Server error

    post:
      tags: [profile]
      summary: Update profile photo
      description: Uploads or replaces the user profile photo. Photo is resized to 200x200 px and converted to JPEG.
      operationId: setMyPhoto
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PhotoUpload'
      responses:
        "200":
          description: Photo updated successfully
          content:
            application/json:
              schema:
                type: object
                description: Empty response on success
        "400":
          description: Invalid request (missing username, invalid file format, or file processing error)
        "401":
          description: Unauthorized - invalid or missing token
        "500":
          description: Server error
...
  /users:
    get:
      tags: [profile]
      summary: Search users
      description: Retrieves a list of all users in the system.
      operationId: searchUsers
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: List of all users
          content:
            application/json:
              schema:
                type: array
                minItems: 0
                maxItems: 1000
                items:
                  type: object
                  description: User information
                  properties:
                    username:
                      type: string
                      minLength: 1
                      maxLength: 64
                      pattern: "^[a-zA-Z0-9_-]+$"
                      description: Username of the user
        "401":
          description: Unauthorized
        "500":
          description: Server error
  /users/{username}/photo:
    get:
      tags: [profile]
      summary: Get user profile photo
      description: Retrieves the profile photo of a specific user.
      operationId: getUserProfilePhoto
      security: [{ bearerAuth: [] }]
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 64
            pattern: "^[a-zA-Z0-9_-]+$"
            example: "Karol"
      responses:
        "200":
          description: User profile photo retrieved successfully
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
        "401":
          description: Unauthorized
        "404":
          description: User or photo not found
        "500":
          description: Server error
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    PhotoUpload:
      type: object
      description: Photo upload payload
      required: [photo, username]
      properties:
        photo:
          type: string
          description: Binary image file
          format: binary
        username:
          type: string
          description: Username of the user uploading the photo
          minLength: 1
          maxLength: 64
          pattern: "^[a-zA-Z0-9_-]+$"
          example: "Karol"
...